Fix Room entity
